Add files via upload
CoinApp.java asks the user to input some number of cents (less than 100). The program outputs the number of quarters, dimes, nickels and pennies needed to add up to that amount.

The Employee.java class includes three instance variables — a first name (type String), a last name (type String) and a monthly salary (double). Provides a constructor that initializes the three instance variables and a set and a get method for each instance variable. If the monthly salary is not positive, the value is not set. Includes a toString method that returns a one-line description of the employee.

EmployeeTest.java demonstrates class Employee’s capabilities by creating two Employee objects and displaying each object’s yearly salary, then gives each Employee a 10% raise and displays each Employee’s yearly salary again.

The Dog.java class contains instance data that represents the dog’s name and age, defines the Dog constructor to accept and initialize instance data, includes getter and setter methods for the name and age, a method to compute and return the age of the dog in “person years” (seven times the dog’s age), and a toString method that returns a one-line description of the dog.

The main method of the driver class Kennel.java instantiates and updates several Dog objects.